requirements.txt
#
# Python library requirements for netlab
#
Jinja2 >= 2.11.0
PyYAML >= 5.0.0
netaddr >= 0.8.0
python-box >= 5.4.1,<6.0
importlib_resources
typing-extensions
schema >= 0.7.5
